Send and receive long, not int, now that bufsz is a long
This triggered problems with very large reallocs on lobo, because the
high bits of inbytes were non-zero.  I'm a little worried that this
might be causing problems for, e.g., Carola on saguaro, but I guess
inbytes happens to be zero in her binary.  Need to communicate this
change.
(cherry picked from commit 8d84f7741a71782e4625a90bd9f8993e9fbff946)

+++ libsw/mpmy_gather.c
@@ -201,11 +201,11 @@ MPMY_NGather(const void *sendbuf, int count, MPMY_Datatype type,
 	if (sendproc >= 0 && sendproc < nproc) {
 	    if (procnum & (1 << chan)) {
 		Msgf(("NGather: to %d, bufsz %d\n", sendproc, bufsz));
-		MPMY_send(&bufsz, sizeof(int), sendproc, NGATHER_TAG+chan);
+		MPMY_send(&bufsz, sizeof(long), sendproc, NGATHER_TAG+chan);
 		MPMY_send(buf, bufsz, sendproc, NGATHER_TAG+chan);
 		break;
 	    } else {
-		MPMY_recvn(&inbytes, sizeof(int), sendproc, NGATHER_TAG+chan);
+		MPMY_recvn(&inbytes, sizeof(long), sendproc, NGATHER_TAG+chan);
 		Msgf(("NGather: from %d, inbytes %d\n", sendproc, inbytes));
 		buf = Realloc(buf, bufsz+inbytes);
 		MPMY_recvn(bufsz+(char*)buf, inbytes, sendproc, NGATHER_TAG+chan);
